pythonFacialRec/requirements.txt
cryptography
face_recognition
opencv-python
mediapipe
numpy
scikit-learn
watchdog
aiohttp
psycopg2
python-dotenv
